refactor(tui): complete event_bus and lru_cache module migration
Delete old event.rs and preview_cache.rs that were coexisting with
their replacements (event_bus.rs and lru_cache/). Update all import
paths and fix API incompatibilities (invalidate takes &String not
&Path, use new_preview_cache() instead of PreviewCache::default()).
