+++ src/tui/cleanup_app.rs
@@ -16,8 +16,8 @@ use ratatui::{
 };
 
 use super::app::App;
-use super::event::Event;
-use super::preview_cache::PreviewCache;
+use super::event_bus::Event;
+use super::lru_cache::{new_preview_cache, PreviewCache};
 use super::widgets::{FileExplorer, FileExplorerWidget, FileItem};
 use crate::theme::current_theme;
 
@@ -87,7 +87,7 @@ impl CleanupApp {
             available_agents,
             status_message: None,
             files_deleted: false,
-            preview_cache: PreviewCache::default(),
+            preview_cache: new_preview_cache(),
         })
     }
